removed and fixed yesterday's todos

+++ Shared/interfaces/IBucketListData.cs
@@ -7,15 +7,15 @@ namespace Shared.interfaces
     {
         IList<Shared.dto.BucketListItem> GetBucketList(string userName);
         void UpsertBucketListItem(Shared.dto.BucketListItem bucketListItem, string userName);
-        void DeleteBucketListItem(int bucketListItemDbId);
+        void DeleteBucketListItem(long bucketListItemDbId);
         void LogMsg(string msg);
         IList<SystemStatistic> GetSystemStatistics();
         IList<SystemBuildStatistic> GetSystemBuildStatistics();
 
-        void AddToken(int userId, string token);
-        User GetUser(int id);
+        void AddToken(long userId, string token);
+        User GetUser(long id);
         User GetUser(string userName);
-        int AddUser(User user);
-        void DeleteUser(int userId);
+        long AddUser(User user);
+        void DeleteUser(long userId);
     }
 }
